push descriptor outside of device specific impl + desc class w/ prec and dom
